+++ examples/test/lines.py
@@ -1,8 +1,7 @@
 from easygraphics import *
 
 init_graph(600, 600)
-points = [50, 50, 550, 350, 50, 150, 550, 450, 50, 250, 550, 550]
-draw_lines(points)
+draw_lines(50, 50, 550, 350, 50, 150, 550, 450, 50, 250, 550, 550)
 pause()
 close_graph()