+++ examples/tutorials/04_polygon.py
@@ -7,7 +7,6 @@ if __name__ == "__main__":
     init_graph(400, 300)
     set_color(Color.DARK_BLUE)
     set_fill_color(Color.LIGHT_MAGENTA)
-    points = [50, 50, 350, 250, 50, 150]
-    draw_polygon(points)
+    draw_polygon(50, 50, 350, 250, 50, 150)
     pause()
     close_graph()