+++ examples/tutorials/12_animation.py
@@ -20,7 +20,7 @@ if __name__ == "__main__":
     set_composition_mode(CompositionMode.SOURCE)
     set_background_color(Color.TRANSPARENT)
     set_fill_color("white")
-    draw_polygon([0, 0, 0, 60, 160, 60, 160, 40, 125, 20, 110, 0])
+    draw_polygon(0, 0, 0, 60, 160, 60, 160, 40, 125, 20, 110, 0)
     set_fill_color("darkgray")
     draw_circle(35, 60, 20)
     draw_circle(120, 60, 20)