+++ examples/test/reinit.py
@@ -2,8 +2,11 @@ from easygraphics import *
 import time
 
 init_graph()
+pause()
 time.sleep(1)
+print("lala")
 close_graph()
+print("haha")
 time.sleep(1)
 init_graph()
 pause()